Refactor workspace context usage to improve structure and introduce ThemeContext
